+++ atlasdb-commons/src/main/java/com/palantir/async/initializer/LambdaCallback.java
@@ -16,21 +16,67 @@
 
 package com.palantir.async.initializer;
 
+import java.util.function.BiConsumer;
 import java.util.function.Consumer;
 
-public final class LambdaCallback<R> extends Callback<R> {
-    private final Consumer<R> consumer;
+import com.palantir.common.base.Throwables;
 
-    private LambdaCallback(Consumer<R> consumer) {
-        this.consumer = consumer;
+/**
+ * Convenience class for creating callbacks using lambda expressions.
+ */
+public final class LambdaCallback<R> extends Callback<R> {
+    private static final BiConsumer<Object, Throwable> FAIL = (ignore, throwable) -> {
+        throw Throwables.rewrapAndThrowUncheckedException(throwable);
+    };
+
+    private final Consumer<R> initialize;
+    private final BiConsumer<R, Throwable> onInitializationFailure;
+
+    private LambdaCallback(Consumer<R> initialize, BiConsumer<R, Throwable> onInitializationFailure) {
+        this.initialize = initialize;
+        this.onInitializationFailure = onInitializationFailure;
     }
 
-    public static <R> Callback<R> of(Consumer<R> consumer) {
-        return new LambdaCallback<>(consumer);
+    /**
+     * Creates a callback that will attempt to initialize once. If initialize throws, onInitFailureCleanup is called
+     * before wrapping and throwing a RuntimeException.
+     * @param initialize initialization method.
+     * @param onInitFailureCleanup cleanup to be done in case initialization throws. If this method also throws, it will
+     * bubble up to the caller.
+     * @return the desired Callback object.
+     */
+    public static <R> Callback<R> singleAttempt(Consumer<R> initialize, BiConsumer<R, Throwable> onInitFailureCleanup) {
+        return new LambdaCallback<>(initialize, onInitFailureCleanup.andThen(FAIL));
+    }
+
+    /**
+     * Creates a callback that will retry initialization until cleanup also throws. If initialize throws,
+     * onInitFailureCleanup is called before calling initialize again.
+     * @param initialize initialization method.
+     * @param onInitFailureCleanup cleanup to be done in case initialization throws. If this method also throws, no more
+     * retries will be attampted and it will bubble up to the caller.
+     * @return the desired Callback object.
+     */
+    public static <R> Callback<R> retrying(Consumer<R> initialize, BiConsumer<R, Throwable> onInitFailureCleanup) {
+        return new LambdaCallback<>(initialize, onInitFailureCleanup);
+    }
+
+    /**
+     * Simple version of {@link #singleAttempt(Consumer, BiConsumer)} when no cleanup is necessary on failure.
+     * @param initialize initialization method.
+     * @return the desired Callback object.
+     */
+    public static <R> Callback<R> of(Consumer<R> initialize) {
+        return singleAttempt(initialize, (no, cleanup) -> { });
     }
 
     @Override
     public void init(R resource) {
-        consumer.accept(resource);
+        initialize.accept(resource);
+    }
+
+    @Override
+    public void cleanup(R resource, Throwable initFailure) {
+        onInitializationFailure.accept(resource, initFailure);
     }
 }